+++ tests/Unit/Filesystem/FinderTest.php
@@ -8,15 +8,21 @@ use PabloK\SupercacheBundle\Exceptions\EmptyPathException;
 use PabloK\SupercacheBundle\Filesystem\Finder;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Filesystem\Filesystem;
+use voku\helper\HtmlMin;
 
 class FinderTest extends TestCase
 {
     public function testUnableToPassEmptyCacheDir()
     {
         $mockFilesystem = $this->createMock(Filesystem::class);
+        $mockHtmlMin = $this->createMock(HtmlMin::class);
 
         $this->expectException(EmptyPathException::class);
 
-        new Finder('', $mockFilesystem);
+        new Finder(
+            '',
+            $mockFilesystem,
+            $mockHtmlMin
+        );
     }
 }